Module fastsim_core::pyo3imports

Modules

asyncio

Exceptions defined in Python’s asyncio module.

socket

Exceptions defined in Python’s socket module.

Macros

wrap_pyfunction

Wraps a Rust function annotated with #[pyfunction].

Structs

Py

A GIL-independent reference to an object allocated on the Python heap.

PyAny

Represents any Python object.

PyArithmeticError

Represents Python’s ArithmeticError exception.

PyAssertionError

Represents Python’s AssertionError exception.

PyAttributeError

Represents Python’s AttributeError exception.

PyBaseException

Represents Python’s BaseException exception.

PyBlockingIOError

Represents Python’s BlockingIOError exception.

PyBrokenPipeError

Represents Python’s BrokenPipeError exception.

PyBufferError

Represents Python’s BufferError exception.

PyBytes

Represents a Python bytes object.

PyBytesWarning

Represents Python’s BytesWarning exception.

PyCell

A container type for (mutably) accessing [PyClass] values

PyChildProcessError

Represents Python’s ChildProcessError exception.

PyClassInitializer

Initializer for our #[pyclass] system.

PyConnectionAbortedError

Represents Python’s ConnectionAbortedError exception.

PyConnectionError

Represents Python’s ConnectionError exception.

PyConnectionRefusedError

Represents Python’s ConnectionRefusedError exception.

PyConnectionResetError

Represents Python’s ConnectionResetError exception.

PyDeprecationWarning

Represents Python’s DeprecationWarning exception.

PyDict

Represents a Python dict.

PyEOFError

Represents Python’s EOFError exception.

PyEncodingWarning

Represents Python’s EncodingWarning exception.

PyEnvironmentError

Represents Python’s EnvironmentError exception.

PyErr

Represents a Python exception.

PyException

Represents Python’s Exception exception.

PyFileExistsError

Represents Python’s FileExistsError exception.

PyFileNotFoundError

Represents Python’s FileNotFoundError exception.

PyFloatingPointError

Represents Python’s FloatingPointError exception.

PyFutureWarning

Represents Python’s FutureWarning exception.

PyGeneratorExit

Represents Python’s GeneratorExit exception.

PyIOError

Represents Python’s IOError exception.

PyImportError

Represents Python’s ImportError exception.

PyImportWarning

Represents Python’s ImportWarning exception.

PyIndexError

Represents Python’s IndexError exception.

PyInterruptedError

Represents Python’s InterruptedError exception.

PyIsADirectoryError

Represents Python’s IsADirectoryError exception.

PyKeyError

Represents Python’s KeyError exception.

PyKeyboardInterrupt

Represents Python’s KeyboardInterrupt exception.

PyLookupError

Represents Python’s LookupError exception.

PyMemoryError

Represents Python’s MemoryError exception.

PyModule

Represents a Python module object.

PyModuleNotFoundError

Represents Python’s ModuleNotFoundError exception.

PyNameError

Represents Python’s NameError exception.

PyNotADirectoryError

Represents Python’s NotADirectoryError exception.

PyNotImplementedError

Represents Python’s NotImplementedError exception.

PyOSError

Represents Python’s OSError exception.

PyOverflowError

Represents Python’s OverflowError exception.

PyPendingDeprecationWarning

Represents Python’s PendingDeprecationWarning exception.

PyPermissionError

Represents Python’s PermissionError exception.

PyProcessLookupError

Represents Python’s ProcessLookupError exception.

PyRecursionError

Represents Python’s RecursionError exception.

PyRef

A wrapper type for an immutably borrowed value from a PyCell<T>.

PyRefMut

A wrapper type for a mutably borrowed value from aPyCell<T>.

PyReferenceError

Represents Python’s ReferenceError exception.

PyResourceWarning

Represents Python’s ResourceWarning exception.

PyRuntimeError

Represents Python’s RuntimeError exception.

PyRuntimeWarning

Represents Python’s RuntimeWarning exception.

PyStopAsyncIteration

Represents Python’s StopAsyncIteration exception.

PyStopIteration

Represents Python’s StopIteration exception.

PySyntaxError

Represents Python’s SyntaxError exception.

PySyntaxWarning

Represents Python’s SyntaxWarning exception.

PySystemError

Represents Python’s SystemError exception.

PySystemExit

Represents Python’s SystemExit exception.

PyTimeoutError

Represents Python’s TimeoutError exception.

PyType

Represents a reference to a Python type object.

PyTypeError

Represents Python’s TypeError exception.

PyUnboundLocalError

Represents Python’s UnboundLocalError exception.

PyUnicodeDecodeError

Represents Python’s UnicodeDecodeError exception.

PyUnicodeEncodeError

Represents Python’s UnicodeEncodeError exception.

PyUnicodeError

Represents Python’s UnicodeError exception.

PyUnicodeTranslateError

Represents Python’s UnicodeTranslateError exception.

PyUnicodeWarning

Represents Python’s UnicodeWarning exception.

PyUserWarning

Represents Python’s UserWarning exception.

PyValueError

Represents Python’s ValueError exception.

PyWarning

Represents Python’s Warning exception.

PyZeroDivisionError

Represents Python’s ZeroDivisionError exception.

Python

A marker token that represents holding the GIL.

Traits

FromPyObject

Extract a type from a Python object.

IntoPy

Defines a conversion from a Rust type to a Python object.

IntoPyPointer

Returns an owned pointer to a Python object.

PyTryFrom

Trait implemented by Python object types that allow a checked downcast. If T implements PyTryFrom, we can convert &PyAny to &T.

PyTryInto

Trait implemented by Python object types that allow a checked downcast. This trait is similar to std::convert::TryInto

ToPyObject

Conversion trait that allows various objects to be converted into PyObject.

Type Aliases

PyObject

A commonly-used alias for Py<PyAny>.

PyResult

Represents the result of a Python call.

Attribute Macros

pyclass

pyfunction

A proc macro used to expose Rust functions to Python.

pymethods

A proc macro used to expose methods to Python.

pymodule

A proc macro used to implement Python modules.

Derive Macros

FromPyObject