Struct fastsim_core::simdrive::RustSimDriveParams

pub struct RustSimDriveParams {

    pub favor_grade_accuracy: bool,
    pub missed_trace_correction: bool,
    pub max_time_dilation: f64,
    pub min_time_dilation: f64,
    pub time_dilation_tol: f64,
    pub max_trace_miss_iters: u32,
    pub trace_miss_speed_mps_tol: f64,
    pub trace_miss_time_tol: f64,
    pub trace_miss_dist_tol: f64,
    pub sim_count_max: usize,
    pub newton_gain: f64,
    pub newton_max_iter: u32,
    pub newton_xtol: f64,
    pub energy_audit_error_tol: f64,
    pub coast_allow: bool,
    pub coast_allow_passing: bool,
    pub coast_max_speed_m_per_s: f64,
    pub coast_brake_accel_m_per_s2: f64,
    pub coast_brake_start_speed_m_per_s: f64,
    pub coast_start_speed_m_per_s: f64,
    pub coast_time_horizon_for_adjustment_s: f64,
    pub idm_allow: bool,
    pub idm_v_desired_m_per_s: f64,
    pub idm_dt_headway_s: f64,
    pub idm_minimum_gap_m: f64,
    pub idm_delta: f64,
    pub idm_accel_m_per_s2: f64,
    pub idm_decel_m_per_s2: f64,
    pub idm_v_desired_in_m_per_s_by_distance_m: Option<Vec<(f64, f64)>>,
    pub max_epa_adj: f64,
    pub orphaned: bool,

}

Struct containing time trace data

Fields
favor_grade_accuracy: bool

if true, accuracy will be favored over performance for grade per step estimates Specifically, for performance, grade for a step will be assumed to be the grade looked up at step start distance. For accuracy, the actual elevations will be used. This distinciton only makes a difference for CAV maneuvers.

missed_trace_correction: bool

if true, missed trace correction is active, default = False. If missed trace correction is active, time step will be “dilated” to be long enough for vehicle to “catch up” with trace.

max_time_dilation: f64

maximum time dilation factor to “catch up” with trace – e.g. 1.0 means 100% increase in step size

min_time_dilation: f64

minimum time dilation margin to let trace “catch up” – e.g. -0.5 means 50% reduction in step size

time_dilation_tol: f64

convergence criteria for time dilation in iterating on time step size to achieve distance parity

max_trace_miss_iters: u32

number of iterations to achieve time dilation correction

trace_miss_speed_mps_tol: f64

threshold for triggering warning log message if vehicle speed deficit [m/s] relative to prescribed speed exceeds this amount

trace_miss_time_tol: f64

threshold for triggering warning log message if achieved elapsed time relative to prescribed elapsed time exceeds this fractional amount

trace_miss_dist_tol: f64

threshold for triggering warning log message if achieved distance relative to prescribed distance exceeds this fractional amount

sim_count_max: usize

max allowable number of HEV SOC iterations

newton_gain: f64

newton solver gain

newton_max_iter: u32

newton solver max iterations

newton_xtol: f64

newton solver tolerance

energy_audit_error_tol: f64

tolerance for energy audit error warning, i.e. 0.1%

coast_allow: bool

if true, coasting to stops are allowed

coast_allow_passing: bool

if true, coasting vehicle can eclipse the shadow trace (i.e., reference vehicle in front)

coast_max_speed_m_per_s: f64

maximum allowable speed under coast (m/s)

coast_brake_accel_m_per_s2: f64

acceleration assumed during braking for coast maneuvers (m/s2). note: should be negative

coast_brake_start_speed_m_per_s: f64

speed when friction braking will initiate during coasting maneuvers (m/s)

coast_start_speed_m_per_s: f64

initiates coast when vehicle hits this speed if > 0; this is mainly for forceing coasting to initiate for testing. (m/s)

coast_time_horizon_for_adjustment_s: f64

“look-ahead” time for speed changes to be considered to feature coasting to hit a given stopping distance mark (s)

idm_allow: bool

if true, initiates the IDM - Intelligent Driver Model, Adaptive Cruise Control version

idm_v_desired_m_per_s: f64

IDM algorithm: desired speed (m/s)

idm_dt_headway_s: f64

IDM algorithm: headway time desired to vehicle in front (s)

idm_minimum_gap_m: f64

IDM algorithm: minimum desired gap between vehicle and lead vehicle (m)

idm_delta: f64

IDM algorithm: delta parameter

idm_accel_m_per_s2: f64

IDM algorithm: acceleration parameter

idm_decel_m_per_s2: f64

IDM algorithm: deceleration parameter

idm_v_desired_in_m_per_s_by_distance_m: Option<Vec<(f64, f64)>>

IDM algorithm: a way to specify desired speed by course distance traveled. Can simulate changing speed limits over a driving cycle optional list of (distance (m), desired speed (m/s))

max_epa_adj: f64

EPA fuel economy adjustment parameters; maximum EPA adjustment factor

orphaned: bool
Implementations

impl RustSimDriveParams

pub fn __getnewargs__(&self)

pub fn get_favor_grade_accuracy(&self) -> bool

pub fn set_favor_grade_accuracy(&mut self, new_value: bool) -> Result<()>

pub fn get_missed_trace_correction(&self) -> bool

pub fn set_missed_trace_correction(&mut self, new_value: bool) -> Result<()>

pub fn get_max_time_dilation(&self) -> f64

pub fn set_max_time_dilation(&mut self, new_value: f64) -> Result<()>

pub fn get_min_time_dilation(&self) -> f64

pub fn set_min_time_dilation(&mut self, new_value: f64) -> Result<()>

pub fn get_time_dilation_tol(&self) -> f64

pub fn set_time_dilation_tol(&mut self, new_value: f64) -> Result<()>

pub fn get_max_trace_miss_iters(&self) -> u32

pub fn set_max_trace_miss_iters(&mut self, new_value: u32) -> Result<()>

pub fn get_trace_miss_speed_mps_tol(&self) -> f64

pub fn set_trace_miss_speed_mps_tol(&mut self, new_value: f64) -> Result<()>

pub fn get_trace_miss_time_tol(&self) -> f64

pub fn set_trace_miss_time_tol(&mut self, new_value: f64) -> Result<()>

pub fn get_trace_miss_dist_tol(&self) -> f64

pub fn set_trace_miss_dist_tol(&mut self, new_value: f64) -> Result<()>

pub fn get_sim_count_max(&self) -> usize

pub fn set_sim_count_max(&mut self, new_value: usize) -> Result<()>

pub fn get_newton_gain(&self) -> f64

pub fn set_newton_gain(&mut self, new_value: f64) -> Result<()>

pub fn get_newton_max_iter(&self) -> u32

pub fn set_newton_max_iter(&mut self, new_value: u32) -> Result<()>

pub fn get_newton_xtol(&self) -> f64

pub fn set_newton_xtol(&mut self, new_value: f64) -> Result<()>

pub fn get_energy_audit_error_tol(&self) -> f64

pub fn set_energy_audit_error_tol(&mut self, new_value: f64) -> Result<()>

pub fn get_coast_allow(&self) -> bool

pub fn set_coast_allow(&mut self, new_value: bool) -> Result<()>

pub fn get_coast_allow_passing(&self) -> bool

pub fn set_coast_allow_passing(&mut self, new_value: bool) -> Result<()>

pub fn get_coast_max_speed_m_per_s(&self) -> f64

pub fn set_coast_max_speed_m_per_s(&mut self, new_value: f64) -> Result<()>

pub fn get_coast_brake_accel_m_per_s2(&self) -> f64

pub fn set_coast_brake_accel_m_per_s2(&mut self, new_value: f64) -> Result<()>

pub fn get_coast_brake_start_speed_m_per_s(&self) -> f64

pub fn set_coast_brake_start_speed_m_per_s( &mut self, new_value: f64, ) -> Result<()>

pub fn get_coast_start_speed_m_per_s(&self) -> f64

pub fn set_coast_start_speed_m_per_s(&mut self, new_value: f64) -> Result<()>

pub fn get_coast_time_horizon_for_adjustment_s(&self) -> f64

pub fn set_coast_time_horizon_for_adjustment_s( &mut self, new_value: f64, ) -> Result<()>

pub fn get_idm_allow(&self) -> bool

pub fn set_idm_allow(&mut self, new_value: bool) -> Result<()>

pub fn get_idm_v_desired_m_per_s(&self) -> f64

pub fn set_idm_v_desired_m_per_s(&mut self, new_value: f64) -> Result<()>

pub fn get_idm_dt_headway_s(&self) -> f64

pub fn set_idm_dt_headway_s(&mut self, new_value: f64) -> Result<()>

pub fn get_idm_minimum_gap_m(&self) -> f64

pub fn set_idm_minimum_gap_m(&mut self, new_value: f64) -> Result<()>

pub fn get_idm_delta(&self) -> f64

pub fn set_idm_delta(&mut self, new_value: f64) -> Result<()>

pub fn get_idm_accel_m_per_s2(&self) -> f64

pub fn set_idm_accel_m_per_s2(&mut self, new_value: f64) -> Result<()>

pub fn get_idm_decel_m_per_s2(&self) -> f64

pub fn set_idm_decel_m_per_s2(&mut self, new_value: f64) -> Result<()>

pub fn get_idm_v_desired_in_m_per_s_by_distance_m( &self, ) -> Option<Vec<(f64, f64)>>

pub fn set_idm_v_desired_in_m_per_s_by_distance_m( &mut self, new_value: Option<Vec<(f64, f64)>>, ) -> Result<()>

pub fn get_max_epa_adj(&self) -> f64

pub fn set_max_epa_adj(&mut self, new_value: f64) -> Result<()>

pub fn get_orphaned(&self) -> bool

pub fn reset_orphaned(&mut self)

Reset the orphaned flag to false.

pub fn copy(&self) -> Self

pub fn __copy__(&self) -> Self

pub fn __deepcopy__(&self, _memo: &PyDict) -> Self

pub fn from_resource_py( filepath: &PyAny, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a resource file packaged with the fastsim-core crate

Arguments:

filepath: str | pathlib.Path - Filepath, relative to the top of the resources folder (excluding any relevant prefix), from which to read the object

pub fn to_file_py(&self, filepath: &PyAny) -> PyResult<()>

Write (serialize) an object to a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS. Creates a new file if it does not already exist, otherwise truncates the existing file.

Arguments

filepath: str | pathlib.Path - The filepath at which to write the object

pub fn from_file_py(filepath: &PyAny, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS.

Arguments:

filepath: str | pathlib.Path - The filepath from which to read the object

pub fn to_str_py(&self, format: &str) -> PyResult<String>

Write (serialize) an object into a string

Arguments:

format: str - The target format, any of those listed in ACCEPTED_STR_FORMATS

pub fn from_str_py( contents: &str, format: &str, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a string

Arguments:

contents: str - The string containing the object data

format: str - The source format, any of those listed in ACCEPTED_STR_FORMATS

pub fn to_json_py(&self) -> PyResult<String>

Write (serialize) an object to a JSON string

pub fn from_json_py(json_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object to a JSON string

Arguments

json_str: str - JSON-formatted string to deserialize from

pub fn to_yaml_py(&self) -> PyResult<String>

Write (serialize) an object to a YAML string

pub fn from_yaml_py(yaml_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a YAML string

Arguments

yaml_str: str - YAML-formatted string to deserialize from