Function fastsim_core::utils::multilinear

pub fn multilinear(
    point: &[f64],
    grid: &[Vec<f64>],
    values: &ArrayD<f64>,
) -> Result<f64>

Multilinear interpolation function, accepting any dimensionality N.

Arguments

point - An N-length array representing the interpolation point coordinates in each dimension

grid - A grid containing the coordinates for each dimension, i.e. [[0.0, 1.0], [-0.5, 1.5]] indicates x0 = 0.0, x1 = 1.0, y0 = -0.5, y1 = 1.5

values - An N-dimensional ndarray::ArrayD containing the values at given grid coordinates

Errors

This function returns an [InterpolationError] if any of the validation checks from [validate_inputs] fail, or if any values surrounding supplied point are NaN.

Examples

1D Example

use ndarray::prelude::*;
use fastsim_core::utils::multilinear;

let grid = [vec![0.0, 1.0, 4.0]];
let values = array![0.0, 2.0, 4.45].into_dyn();

let point_a = [0.82];
assert_eq!(multilinear(&point_a, &grid, &values).unwrap(), 1.64);
let point_b = [2.98];
assert_eq!(multilinear(&point_b, &grid, &values).unwrap(), 3.617);
let point_c = [grid[0][2]]; // returns value at x2
assert_eq!(multilinear(&point_c, &grid, &values).unwrap(), values[2]);

2D Example

use ndarray::prelude::*;
use fastsim_core::utils::multilinear;

let grid = [
    vec![0.0, 1.0, 2.0], // x0, x1, x2
    vec![0.0, 1.0, 2.0], // y0, y1, y2
];
let values = array![
    [0.0, 2.0, 1.9], // (x0, y0), (x0, y1), (x0, y2)
    [2.0, 4.0, 3.1], // (x1, y0), (x1, y1), (x1, y2)
    [5.0, 0.0, 1.4], // (x2, y0), (x2, y1), (x2, y2)
]
.into_dyn();

let point_a = [0.5, 0.5];
assert_eq!(multilinear(&point_a, &grid, &values).unwrap(), 2.0);
let point_b = [1.52, 0.36];
assert_eq!(multilinear(&point_b, &grid, &values).unwrap(), 2.9696);
let point_c = [grid[0][2], grid[1][1]]; // returns value at (x2, y1)
assert_eq!(
    multilinear(&point_c, &grid, &values).unwrap(),
    values[[2, 1]]
);

2D Example with non-uniform dimension sizes

use ndarray::prelude::*;
use fastsim_core::utils::multilinear;

let grid = [
    vec![0.0, 1.0, 2.0], // x0, x1, x2
    vec![0.0, 1.0], // y0, y1
];
let values = array![
    [0.0, 2.0], // f(x0, y0), f(x0, y1)
    [2.0, 4.0], // f(x1, y0), f(x1, y1)
    [5.0, 0.0], // f(x2, y0), f(x2, y1)
]
.into_dyn();

let point_a = [0.5, 0.5];
assert_eq!(multilinear(&point_a, &grid, &values).unwrap(), 2.0);
let point_b = [1.52, 0.36];
assert_eq!(multilinear(&point_b, &grid, &values).unwrap(), 2.9696);
let point_c = [grid[0][2], grid[1][1]]; // returns value at (x2, y1)
assert_eq!(
    multilinear(&point_c, &grid, &values).unwrap(),
    values[[2, 1]]
);

3D Example

use ndarray::prelude::*;
use fastsim_core::utils::multilinear;

let grid = [
    vec![0.0, 1.0, 2.0], // x0, x1, x2
    vec![0.0, 1.0, 2.0], // y0, y1, y2
    vec![0.0, 1.0, 2.0], // z0, z1, z2
];
let values = array![
    [
        [0.0, 1.5, 3.0], // (x0, y0, z0), (x0, y0, z1), (x0, y0, z2)
        [2.0, 0.5, 1.4], // (x0, y1, z0), (x0, y1, z1), (x0, y1, z2)
        [1.9, 5.3, 2.2], // (x0, y2, z0), (x0, y0, z1), (x0, y2, z2)
    ],
    [
        [2.0, 5.1, 1.1], // (x1, y0, z0), (x1, y0, z1), (x1, y0, z2)
        [4.0, 1.0, 0.5], // (x1, y1, z0), (x1, y1, z1), (x1, y1, z2)
        [3.1, 0.9, 1.2], // (x1, y2, z0), (x1, y2, z1), (x1, y2, z2)
    ],
    [
        [5.0, 0.2, 5.1], // (x2, y0, z0), (x2, y0, z1), (x2, y0, z2)
        [0.7, 0.1, 3.2], // (x2, y1, z0), (x2, y1, z1), (x2, y1, z2)
        [1.4, 1.1, 0.0], // (x2, y2, z0), (x2, y2, z1), (x2, y2, z2)
    ],
]
.into_dyn();

let point_a = [0.5, 0.5, 0.5];
assert_eq!(multilinear(&point_a, &grid, &values).unwrap(), 2.0125);
let point_b = [1.52, 0.36, 0.5];
assert_eq!(multilinear(&point_b, &grid, &values).unwrap(), 2.46272);
let point_c = [grid[0][2], grid[1][1], grid[2][0]]; // returns value at (x2, y1, z0)
assert_eq!(
    multilinear(&point_c, &grid, &values).unwrap(),
    values[[2, 1, 0]]
);