Module fastsim_core::vehicle_import

Structs

EmissionsInfoFE

Struct containing emissions test results from fueleconomy.gov

EmissionsListFE

Struct containing list of emissions tests from fueleconomy.gov

OptionFE

Struct containing transmission and id of a vehicle option from fueleconomy.gov

OtherVehicleInputs

VehicleDataEPA

Struct containing vehicle data from EPA database

VehicleDataFE

Struct containing vehicle data from fueleconomy.gov

VehicleInputRecord

Functions

get_default_cache_url

get_options_for_year_make_model

Gets options from fueleconomy.gov for the given vehicle year, make, and model

get_vehicle_data_for_id

import_all_vehicles

Import All Vehicles for the given Year, Make, and Model and supplied other inputs

import_all_vehicles_from_record

import_and_save_all_vehicles

import_and_save_all_vehicles_from_file

Import and Save All Vehicles Specified via Input File

vehicle_import_by_id_and_year

Creates RustVehicle for the given vehicle using data from fueleconomy.gov and EPA databases The created RustVehicle is also written as a yaml file