Struct fastsim_core::vehicle_thermal::VehicleThermal

pub struct VehicleThermal {

    pub fc_c_kj__k: f64,
    pub fc_l: f64,
    pub fc_htc_to_amb_stop: f64,
    pub fc_coeff_from_comb: f64,
    pub tstat_te_sto_deg_c: f64,
    pub tstat_te_delta_deg_c: f64,
    pub rad_eps: f64,
    pub fc_model: FcModelTypes,
    pub ess_c_kj_k: f64,
    pub ess_htc_to_amb: f64,
    pub cabin_hvac_model: CabinHvacModelTypes,
    pub cab_c_kj__k: f64,
    pub cab_l_length: f64,
    pub cab_l_width: f64,
    pub cab_r_to_amb: f64,
    pub cab_htc_to_amb_stop: f64,
    pub exhport_model: ComponentModelTypes,
    pub exhport_ha_to_amb: f64,
    pub exhport_ha_int: f64,
    pub exhport_c_kj__k: f64,
    pub cat_model: ComponentModelTypes,
    pub cat_l: f64,
    pub cat_c_kj__K: f64,
    pub cat_htc_to_amb_stop: f64,
    pub cat_te_lightoff_deg_c: f64,
    pub cat_fc_eta_coeff: f64,
    pub orphaned: bool,

}

Struct for containing vehicle thermal (and related) parameters.

Fields
fc_c_kj__k: f64

parameter fuel converter thermal mass [kJ/K]

fc_l: f64

parameter for engine characteristic length [m] for heat transfer calcs

fc_htc_to_amb_stop: f64

parameter for heat transfer coeff [W / (m ** 2 * K)] from eng to ambient during vehicle stop

fc_coeff_from_comb: f64

coeff. for fraction of combustion heat that goes to fuel converter (engine) thermal mass. Remainder goes to environment (e.g. via tailpipe)

tstat_te_sto_deg_c: f64

parameter for temperature [°C] at which thermostat starts to open

tstat_te_delta_deg_c: f64

temperature delta [°C] over which thermostat is partially open

rad_eps: f64

radiator effectiveness – ratio of active heat rejection from radiator to passive heat rejection

fc_model: FcModelTypes

temperature-dependent efficiency fuel converter (engine or fuel cell) thermal model type

ess_c_kj_k: f64

battery thermal mass [kJ/K]

ess_htc_to_amb: f64

effective (incl. any thermal management system) heat transfer coefficient from battery to ambient

cabin_hvac_model: CabinHvacModelTypes

cabin model internal or external w.r.t. fastsim

cab_c_kj__k: f64

parameter for cabin thermal mass [kJ/K]

cab_l_length: f64

cabin length [m], modeled as a flat plate

cab_l_width: f64

cabin width [m], modeled as a flat plate

cab_r_to_amb: f64

cabin shell thermal resistance [m **2 * K / W]

cab_htc_to_amb_stop: f64

parameter for heat transfer coeff [W / (m ** 2 * K)] from cabin to ambient during vehicle stop

exhport_model: ComponentModelTypes

‘external’ (effectively no model) is default exhaust port model type

exhport_ha_to_amb: f64

thermal conductance [W/K] for heat transfer to ambient

exhport_ha_int: f64

thermal conductance [W/K] for heat transfer from exhaust

exhport_c_kj__k: f64

exhaust port thermal capacitance [kJ/K]

cat_model: ComponentModelTypescat_l: f64

diameter [m] of catalyst as sphere for thermal model

cat_c_kj__K: f64

catalyst thermal capacitance [kJ/K]

cat_htc_to_amb_stop: f64

parameter for heat transfer coeff [W / (m ** 2 * K)] from catalyst to ambient during vehicle stop

cat_te_lightoff_deg_c: f64

lightoff temperature to be used when fc_temp_eff_component == ‘hybrid’

cat_fc_eta_coeff: f64

cat engine efficiency coeff. to be used when fc_temp_eff_component == ‘hybrid’

orphaned: bool

for pyo3 api

Implementations

impl VehicleThermal

pub fn default_py() -> Self

pub fn set_cabin_hvac_model_internal( &mut self, hvac_model: HVACModel, ) -> Result<()>

pub fn get_cabin_model_internal(&self) -> Result<HVACModel>

pub fn set_cabin_hvac_model_external(&mut self) -> Result<()>

pub fn set_fc_model_internal_exponential( &mut self, offset: f64, lag: f64, minimum: f64, fc_temp_eff_component: String, ) -> Result<()>

pub fn set_fc_exp_offset(&mut self, new_offset: f64) -> Result<()>

pub fn set_fc_exp_lag(&mut self, new_lag: f64) -> Result<()>

pub fn set_fc_exp_minimum(&mut self, new_minimum: f64) -> Result<()>

pub fn get_fc_exp_offset(&mut self) -> Result<f64>

pub fn get_fc_exp_lag(&mut self) -> Result<f64>

pub fn get_fc_exp_minimum(&mut self) -> Result<f64>

pub fn get_fc_c_kj__k(&self) -> f64

pub fn set_fc_c_kj__k(&mut self, new_value: f64) -> Result<()>

pub fn get_fc_l(&self) -> f64

pub fn set_fc_l(&mut self, new_value: f64) -> Result<()>

pub fn get_fc_htc_to_amb_stop(&self) -> f64

pub fn set_fc_htc_to_amb_stop(&mut self, new_value: f64) -> Result<()>

pub fn get_fc_coeff_from_comb(&self) -> f64

pub fn set_fc_coeff_from_comb(&mut self, new_value: f64) -> Result<()>

pub fn get_tstat_te_sto_deg_c(&self) -> f64

pub fn set_tstat_te_sto_deg_c(&mut self, new_value: f64) -> Result<()>

pub fn get_tstat_te_delta_deg_c(&self) -> f64

pub fn set_tstat_te_delta_deg_c(&mut self, new_value: f64) -> Result<()>

pub fn get_rad_eps(&self) -> f64

pub fn set_rad_eps(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_c_kj_k(&self) -> f64

pub fn set_ess_c_kj_k(&mut self, new_value: f64) -> Result<()>

pub fn get_ess_htc_to_amb(&self) -> f64

pub fn set_ess_htc_to_amb(&mut self, new_value: f64) -> Result<()>

pub fn get_cab_c_kj__k(&self) -> f64

pub fn set_cab_c_kj__k(&mut self, new_value: f64) -> Result<()>

pub fn get_cab_l_length(&self) -> f64

pub fn set_cab_l_length(&mut self, new_value: f64) -> Result<()>

pub fn get_cab_l_width(&self) -> f64

pub fn set_cab_l_width(&mut self, new_value: f64) -> Result<()>

pub fn get_cab_r_to_amb(&self) -> f64

pub fn set_cab_r_to_amb(&mut self, new_value: f64) -> Result<()>

pub fn get_cab_htc_to_amb_stop(&self) -> f64

pub fn set_cab_htc_to_amb_stop(&mut self, new_value: f64) -> Result<()>

pub fn get_exhport_ha_to_amb(&self) -> f64

pub fn set_exhport_ha_to_amb(&mut self, new_value: f64) -> Result<()>

pub fn get_exhport_ha_int(&self) -> f64

pub fn set_exhport_ha_int(&mut self, new_value: f64) -> Result<()>

pub fn get_exhport_c_kj__k(&self) -> f64

pub fn set_exhport_c_kj__k(&mut self, new_value: f64) -> Result<()>

pub fn get_cat_l(&self) -> f64

pub fn set_cat_l(&mut self, new_value: f64) -> Result<()>

pub fn get_cat_c_kj__K(&self) -> f64

pub fn set_cat_c_kj__K(&mut self, new_value: f64) -> Result<()>

pub fn get_cat_htc_to_amb_stop(&self) -> f64

pub fn set_cat_htc_to_amb_stop(&mut self, new_value: f64) -> Result<()>

pub fn get_cat_te_lightoff_deg_c(&self) -> f64

pub fn set_cat_te_lightoff_deg_c(&mut self, new_value: f64) -> Result<()>

pub fn get_cat_fc_eta_coeff(&self) -> f64

pub fn set_cat_fc_eta_coeff(&mut self, new_value: f64) -> Result<()>

pub fn get_orphaned(&self) -> bool

pub fn reset_orphaned(&mut self)

Reset the orphaned flag to false.

pub fn copy(&self) -> Self

pub fn __copy__(&self) -> Self

pub fn __deepcopy__(&self, _memo: &PyDict) -> Self

pub fn from_resource_py( filepath: &PyAny, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a resource file packaged with the fastsim-core crate

Arguments:

filepath: str | pathlib.Path - Filepath, relative to the top of the resources folder (excluding any relevant prefix), from which to read the object

pub fn to_file_py(&self, filepath: &PyAny) -> PyResult<()>

Write (serialize) an object to a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS. Creates a new file if it does not already exist, otherwise truncates the existing file.

Arguments

filepath: str | pathlib.Path - The filepath at which to write the object

pub fn from_file_py(filepath: &PyAny, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a file. Supported file extensions are listed in ACCEPTED_BYTE_FORMATS.

Arguments:

filepath: str | pathlib.Path - The filepath from which to read the object

pub fn to_str_py(&self, format: &str) -> PyResult<String>

Write (serialize) an object into a string

Arguments:

format: str - The target format, any of those listed in ACCEPTED_STR_FORMATS

pub fn from_str_py( contents: &str, format: &str, skip_init: Option<bool>, ) -> PyResult<Self>

Read (deserialize) an object from a string

Arguments:

contents: str - The string containing the object data

format: str - The source format, any of those listed in ACCEPTED_STR_FORMATS

pub fn to_json_py(&self) -> PyResult<String>

Write (serialize) an object to a JSON string

pub fn from_json_py(json_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object to a JSON string

Arguments

json_str: str - JSON-formatted string to deserialize from

pub fn to_yaml_py(&self) -> PyResult<String>

Write (serialize) an object to a YAML string

pub fn from_yaml_py(yaml_str: &str, skip_init: Option<bool>) -> PyResult<Self>

Read (deserialize) an object from a YAML string

Arguments

yaml_str: str - YAML-formatted string to deserialize from

impl VehicleThermal

pub fn tstat_te_fo_deg_c(&self) -> f64

derived temperature [ºC] at which thermostat is fully open

pub fn fc_area_ext(&self) -> f64

parameter for engine surface area [m**2] for heat transfer calcs

pub fn cat_area_ext(&self) -> f64

parameter for catalyst surface area [m**2] for heat transfer calcs